Generate Reports funcionality added for feedbacks

+++ controller/AdminController.js
@@ -1,12 +1,24 @@
 const Admin = require('../models').Admin;
 const User = require('../models').User;
 
+const checkPermission = require('../permission');
+const fs = require('fs');
+
+const sendConfirmEmail = require('../utils/emailConfirmation');
 const bcrypt = require('bcryptjs');
 const upload = require('../uploader');
+
+const { Op } = require('sequelize');
 
 
 exports.create = [upload.single('photo_path'), async (req, res) => {
     try {
+        checkPermission(req.user.role, 'ADMIN');
+        const emailExist = await User.findOne({ where: { email: req.body.email } });
+        if (emailExist) {
+            return res.status(400).json({ msg: 'This email already exists', success: false })
+        }
+
         let { email, password, role, name, surname, address, phone, birthday } = req.body;
 
         const user = {
@@ -23,6 +35,8 @@ exports.create = [upload.single('photo_path'), async (req, res) => {
             photo_path: req.file.path
         }
 
+
+
         const salt = await bcrypt.genSalt(10);
 
         const hashedPassword = await bcrypt.hash(user.password, salt);
@@ -30,6 +44,8 @@ exports.create = [upload.single('photo_path'), async (req, res) => {
 
         const registeredUser = await User.create(user);
         const registeredAdmin = await registeredUser.createAdmin(admin);
+
+        sendConfirmEmail(registeredUser)
 
         return res.status(200).json({
             success: true,
@@ -45,6 +61,7 @@ exports.create = [upload.single('photo_path'), async (req, res) => {
 
 exports.updateAdmin = async (req, res) => {
     try {
+        checkPermission(req.user.role, 'ADMIN');
         let { name, surname, address, phone, birthday } = req.body;
 
         const updatedAdmin = await Admin.update({
@@ -54,7 +71,12 @@ exports.updateAdmin = async (req, res) => {
             phone,
             birthday,
             include: [
-                { model: User }
+                {
+                    model: User,
+                    attributes: {
+                        exclude: ['password']
+                    }
+                }
             ]
         },
             { where: { id: req.params.id } }
@@ -70,10 +92,16 @@ exports.updateAdmin = async (req, res) => {
 
 exports.findOne = async (req, res) => {
     try {
+        checkPermission(req.user.role, 'ADMIN');
         const admin = await Admin.findOne({
             where: { id: req.params.id },
             include: [
-                { model: User }
+                {
+                    model: User,
+                    attributes: {
+                        exclude: ['password']
+                    }
+                }
             ]
         });
         res.status(200).json({ admin, success: true });
@@ -86,10 +114,14 @@ exports.findOne = async (req, res) => {
 
 exports.findAdmins = async (req, res) => {
     try {
+        // checkPermission(req.user.role, 'ADMIN');
         const admins = await Admin.findAll({
             include: [
                 {
                     model: User,
+                    attributes: {
+                        exclude: ['password']
+                    }
                 },
             ]
         });
@@ -102,12 +134,18 @@ exports.findAdmins = async (req, res) => {
 
 exports.delete = async (req, res) => {
     try {
+        checkPermission(req.user.role, 'ADMIN');
+        const admin = await Admin.findOne({ where: { UserId: req.params.id } })
         const deletedUser = await User.destroy({
-            where: { id: req.params.id }
+            where: {
+                [Op.and]: [{ id: req.params.id }, { role: 'admin' }]
+            },
         });
+        await fs.unlink(admin.photo_path, err => {
+            if (err) throw err;
+        })
         res.status(200).json({ deletedUser, msg: 'Admin has been deleted succesfully', success: true });
     }
-
     catch (err) {
         res.status(400).json({ msg: err, success: false })
     }
